pi-server/requirements.txt
pandas
sqlalchemy
flask
openai
antlr4-python3-runtime==4.12.0
Werkzeug==2.0.3
python-socketio==5.5.2
Flask-SocketIO==5.1.1
parsimonious

